refactor: move Table::prefix to QueryBuilder
