+++ src/Framework/QueryBuilder/Clauses/Join.php
@@ -2,7 +2,7 @@
 
 namespace Give\Framework\QueryBuilder\Clauses;
 
-use Give\Framework\QueryBuilder\Helpers\Table;
+use Give\Framework\QueryBuilder\QueryBuilder;
 use Give\Framework\QueryBuilder\Types\JoinType;
 use InvalidArgumentException;
 
@@ -33,7 +33,7 @@ class Join
      */
     public function __construct($joinType, $table, $alias = null)
     {
-        $this->table    = Table::prefix($table);
+        $this->table    = QueryBuilder::prefixTable($table);
         $this->joinType = $this->getJoinType($joinType);
         $this->alias    = trim($alias);
     }
