+++ src/Framework/QueryBuilder/QueryBuilder.php
@@ -12,6 +12,7 @@ use Give\Framework\QueryBuilder\Concerns\MetaQuery;
 use Give\Framework\QueryBuilder\Concerns\OffsetStatement;
 use Give\Framework\QueryBuilder\Concerns\OrderByStatement;
 use Give\Framework\QueryBuilder\Concerns\SelectStatement;
+use Give\Framework\QueryBuilder\Concerns\TablePrefix;
 use Give\Framework\QueryBuilder\Concerns\UnionOperator;
 use Give\Framework\QueryBuilder\Concerns\WhereClause;
 
@@ -29,6 +30,7 @@ class QueryBuilder
     use OffsetStatement;
     use OrderByStatement;
     use SelectStatement;
+    use TablePrefix;
     use UnionOperator;
     use WhereClause;
